pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
# Disable checks that conflict with Black/Flake8 or are handled better by other tools
disable = [
    "C0114",  # missing-module-docstring (handled by flake8-docstrings if needed)
    "C0115",  # missing-class-docstring (handled by flake8-docstrings if needed)
    "C0116",  # missing-function-docstring (handled by flake8-docstrings if needed)
    "R0903",  # too-few-public-methods (often not relevant)
    "C0103",  # invalid-name (Black handles naming style)
    "W0613",  # unused-argument (often needed for API compatibility)
    "R0801",  # duplicate-code (often false positives)
    "C0301",  # line-too-long (handled by Black and Flake8)
]

[tool.pylint.design]
# Allow reasonable complexity
max-args = 8
max-locals = 20
max-branches = 15

[tool.ruff]
line-length = 120
target-version = "py312"
select = ["E", "F", "I"]
ignore = ["E203", "W503"]

# ---------------------------------------------------------------------------
# PEP 621 project metadata (enables editable install without invoking custom
# CLI parsing in setup.py for build-time commands like egg_info)
# ---------------------------------------------------------------------------
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hidock-next"
version = "0.0.0"
description = "HiDock Next tooling and desktop environment setup"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "HiDock Contributors"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = []

[project.optional-dependencies]
desktop = [
    "customtkinter>=5.2.0; platform_system == 'Windows' or platform_system == 'Linux' or platform_system == 'Darwin'",
    "pyusb>=1.2.1",
]
audio = [
    "numpy>=1.26",
    "pydub>=0.25.1",
    "pygame>=2.5.0",
]
integration = [
    "requests>=2.32.0",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
    "black>=24.4",
    "isort>=5.13",
]

[tool.setuptools]
packages = []  # rely on namespace-less layout; editable install mainly for tests

